requirements.txt
numpy==2.2.6
scipy==1.15.3
soundfile==0.13.1
pyfar==0.7.3
sofar==1.2.2
pyyaml==6.0.2
torch==2.8.0
torchaudio==2.8.0